+++ spec/lib/fitting/cover/json_schema_enum_spec.rb
@@ -10,8 +10,8 @@ RSpec.describe Fitting::Cover::JSONSchemaEnum do
     let(:original) { MultiJson.load(File.read('spec/fixtures/enum/0.json')) }
     let(:first_combination) { MultiJson.load(File.read('spec/fixtures/enum/1.json')) }
     let(:second_combination) { MultiJson.load(File.read('spec/fixtures/enum/2.json')) }
-    let(:first_details) { %w[enum properties.state.ok] }
-    let(:second_details) { %w[enum properties.state.error] }
+    let(:first_details) { %w[enum properties.state.enum.ok] }
+    let(:second_details) { %w[enum properties.state.enum.error] }
 
     it 'returns combinations' do
       expect(subject.combi).to eq([[first_combination, first_details], [second_combination, second_details]])